add show_api_video and downloadable_local_video to episodes table

+++ database/migrations/2022_07_08_210141_add_certificate_code_and_certificate_date_to_transcripts.php
@@ -27,8 +27,13 @@ return new class extends Migration
     public function down()
     {
         Schema::table('transcripts', function (Blueprint $table) {
-            $table->dropColumn('certificate_code');
-            $table->dropColumn('certificate_date');
+            if (
+                Schema::hasColumn('transcripts', 'certificate_code') &&
+                Schema::hasColumn('transcripts', 'certificate_date')
+            ) {
+                $table->dropColumn('certificate_code');
+                $table->dropColumn('certificate_date');
+            }
         });
     }
 };
