+++ database/migrations/2022_07_10_171727_add_softdeletes_to_teachers_table.php
@@ -26,7 +26,9 @@ return new class extends Migration
     public function down()
     {
         Schema::table('teachers', function (Blueprint $table) {
-            $table->dropColumn('deleted_at');
+            if (Schema::hasColumn('teachers', 'deleted_at') ) {
+                $table->dropColumn('deleted_at');
+            }
         });
     }
 };
